+++ src/features/basic/cxpo-order-book/cxpo-order-book.ts
@@ -1,6 +1,7 @@
 import { createFragmentApp } from '@src/utils/vue-fragment-app';
 import OrderBook from './OrderBook.vue';
 import { changeInputValue } from '@src/util';
+import { increaseDefaultBufferSize } from '@src/infrastructure/prun-ui/buffer-sizes';
 
 function onTileReady(tile: PrunTile) {
   if (!tile.parameter) {
@@ -32,6 +33,7 @@ function onTileReady(tile: PrunTile) {
 }
 
 function init() {
+  increaseDefaultBufferSize('CXPO', { width: 60 });
   tiles.observe('CXPO', onTileReady);
 }